Restructured the resources folder
