+++ resources/adminResource.py
@@ -1,18 +1,15 @@
 from flask_restful import Resource, reqparse
-from flask import request
-from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
+from flask import current_app
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from models import Record, db, User
 from datetime import datetime
 from utils import send_email_notification, send_sms_notification
-from werkzeug.utils import secure_filename
-import os
-from app import ALLOWED_EXTENSIONS, MAX_CONTENT_LENGTH, UPLOAD_FOLDER
 
 def is_admin(user_id):
     user = User.query.get(user_id)
     return user and user.is_admin
 
-class AdminRecordResource(Resource):
+class AdminResource(Resource):
     @jwt_required()
     def patch(self, record_id):
         user_id = get_jwt_identity()
